util: centralize configuration reader
ipa-join and ipa-getkeytab have two different config readers. For
ipa-otpd's support for local KDCs we also need to read configuration
files. Centralize the reader based on the ding-libs code.
