+++ client/ipa-getkeytab.c
@@ -37,7 +37,6 @@
 #include <resolv.h>
 #include <sasl/sasl.h>
 #include <popt.h>
-#include <ini_configobj.h>
 #include <openssl/rand.h>
 
 #include "config.h"
@@ -46,7 +45,7 @@
 #include "ipa_asn1.h"
 #include "ipa-client-common.h"
 #include "ipa_ldap.h"
-
+#include "ipa_config.h"
 
 struct srvrec {
     char *host;
@@ -788,87 +787,6 @@ static char *ask_password(krb5_context krbctx, char *prompt1, char *prompt2,
     return password;
 }
 
-struct ipa_config {
-    const char *server_name;
-    const char *domain;
-};
-
-static int config_from_file(struct ini_cfgobj *cfgctx)
-{
-    struct ini_cfgfile *fctx = NULL;
-    char **errors = NULL;
-    int ret;
-
-    ret = ini_config_file_open(IPACONFFILE, 0, &fctx);
-    if (ret) {
-        fprintf(stderr, _("Failed to open config file %s\n"), IPACONFFILE);
-        return ret;
-    }
-
-    ret = ini_config_parse(fctx,
-                           INI_STOP_ON_ANY,
-                           INI_MS_MERGE | INI_MV1S_ALLOW | INI_MV2S_ALLOW,
-                           INI_PARSE_NOWRAP,
-                           cfgctx);
-    if (ret) {
-        fprintf(stderr, _("Failed to parse config file %s\n"), IPACONFFILE);
-        if (ini_config_error_count(cfgctx)) {
-            ini_config_get_errors(cfgctx, &errors);
-            if (errors) {
-                ini_config_print_errors(stderr, errors);
-                ini_config_free_errors(errors);
-            }
-        }
-        ini_config_file_destroy(fctx);
-        return ret;
-    }
-
-    ini_config_file_destroy(fctx);
-    return 0;
-}
-
-static int read_ipa_config(struct ipa_config **ipacfg)
-{
-    struct ini_cfgobj *cfgctx = NULL;
-    struct value_obj *obj = NULL;
-    int ret;
-
-    *ipacfg = calloc(1, sizeof(struct ipa_config));
-    if (!*ipacfg) {
-        return ENOMEM;
-    }
-
-    ret = ini_config_create(&cfgctx);
-    if (ret) {
-        return ENOENT;
-    }
-
-    ret = config_from_file(cfgctx);
-    if (ret) {
-        ini_config_destroy(cfgctx);
-        return EINVAL;
-    }
-
-    ret = ini_get_config_valueobj("global", "server", cfgctx,
-                                  INI_GET_LAST_VALUE, &obj);
-    if (ret != 0 || obj == NULL) {
-        /* if called on an IPA server we need to look for 'host' instead */
-        ret = ini_get_config_valueobj("global", "host", cfgctx,
-                                      INI_GET_LAST_VALUE, &obj);
-    }
-
-    if (ret == 0 && obj != NULL) {
-        (*ipacfg)->server_name = ini_get_string_config_value(obj, &ret);
-    }
-    ret = ini_get_config_valueobj("global", "domain", cfgctx,
-                                  INI_GET_LAST_VALUE, &obj);
-    if (ret == 0 && obj != NULL) {
-        (*ipacfg)->domain = ini_get_string_config_value(obj, &ret);
-    }
-
-    return 0;
-}
-
 static int resolve_ktname(const char *keytab, char **ktname, char **err_msg)
 {
 	char keytab_resolved[PATH_MAX + 1];
@@ -991,6 +909,7 @@ int main(int argc, const char *argv[])
 	int kvno;
 	int i, ret;
 	char *err_msg;
+	struct ipa_config *ipacfg = NULL;
 
     ret = init_gettext();
     if (ret) {
@@ -1086,11 +1005,10 @@ int main(int argc, const char *argv[])
         exit(2);
     }
 
+    ret = ipa_read_config(IPACONFFILE, &ipacfg);
     if (server && (strcasecmp(server, "_srv_") == 0)) {
         struct srvrec *srvrecs, *srv;
-        struct ipa_config *ipacfg = NULL;
 
-        ret = read_ipa_config(&ipacfg);
         if (ret == 0 && ipacfg->domain && verbose) {
             fprintf(stderr, _("DNS discovery for domain %s\n"), ipacfg->domain);
         }
@@ -1122,15 +1040,11 @@ int main(int argc, const char *argv[])
     }
 
     if (!server && !ldap_uri) {
-        struct ipa_config *ipacfg = NULL;
-
-        ret = read_ipa_config(&ipacfg);
-        if (ret == 0) {
+        if (ipacfg->server_name != NULL) {
             server = ipacfg->server_name;
             ipacfg->server_name = NULL;
         }
-        free(ipacfg);
-		if (verbose && server) {
+        if (verbose && server) {
             fprintf(stderr, _("Using server from config %s\n"), server);
         }
         if (!server) {
@@ -1274,4 +1188,9 @@ int main(int argc, const char *argv[])
 			keytab);
 	}
 	exit(0);
+    if (ipacfg) {
+        free(ipacfg->domain);
+        free(ipacfg->server_name);
+        free(ipacfg);
+    }
 }
